refactor cluster-bootstrap/token/util/helpers remove TODO
