+++ cmd/kubeadm/app/phases/bootstraptoken/node/token.go
@@ -24,7 +24,7 @@ import (
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clientset "k8s.io/client-go/kubernetes"
-	bootstraputil "k8s.io/cluster-bootstrap/token/util"
+	tokenutil "k8s.io/cluster-bootstrap/util/tokens"
 
 	"github.com/pkg/errors"
 )
@@ -39,7 +39,7 @@ func UpdateOrCreateTokens(client clientset.Interface, failIfExists bool, tokens 
 
 	for _, token := range tokens {
 
-		secretName := bootstraputil.BootstrapTokenSecretName(token.Token.ID)
+		secretName := tokenutil.BootstrapTokenSecretName(token.Token.ID)
 		secret, err := client.CoreV1().Secrets(metav1.NamespaceSystem).Get(context.TODO(), secretName, metav1.GetOptions{})
 		if secret != nil && err == nil && failIfExists {
 			return errors.Errorf("a token with id %q already exists", token.Token.ID)
